requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
future==1.0.0
iso8601==2.1.0
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.16.0
tk==0.1.0
